requirements.txt
streamlit
requests
python-dotenv
pydub
streamlit-audiorecorder
